refactor shop cart
