--- src/components/cart/Cart.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { useCart } from '../../context/cart';
-
-const cart = () => {
-  return (
-    <div>cart</div>
-  )
-}
-
-export default cart
