+++ src/components/cart/CartShop.jsx
@@ -3,44 +3,9 @@ import { useCart } from '../../context/cart';
 import { API_PAYMENT, API_URL } from '../utils/constants';
 import { AuthContext } from '../../context/AuthContext';
 import { departments, towns } from '../../utils/colombia';
-
-const CartItem = ({ item }) => {
-    const { removeFromCart } = useCart();
-
-    const handleRemove = () => {
-        removeFromCart(item);
-    };
+import CartItem from './CartItem';
 
 
-    return (
-        <div className="justify-between mb-6 rounded-lg bg-white p-6 shadow-md sm:flex sm:justify-start">
-            <img src={item.image} alt={item.name} className="w-45 h-45 rounded-lg sm:w-40" />
-            <div className="sm:ml-4 sm:flex sm:w-full sm:justify-between">
-                <div className="mt-5 sm:mt-0">
-                    <a href={`/productdetail/${item._id}`} className="text-lg font-bold text-gray-900">{item.name}</a>
-                    <p className="mt-2 text-sm text-gray-600">{item.description}</p>
-                </div>
-                <div className="mt-4 flex justify-between sm:space-y-6 sm:mt-0 sm:block sm:space-x-6">
-                    <div className="flex items-center border-gray-100">
-                        <h2 className="text-lg font-bold text-gray-900">Cantidad</h2>
-                        <div className="flex items-center justify-center w-8 h-8 rounded-full bg-gray-100 ml-3">
-                            <p className="text-sm font-bold text-gray-900">{item.quantity}</p>
-                        </div>
-
-
-
-                    </div>
-                    <div className="flex items-center space-x-4">
-                        <p className="text-sm">{item.price} $ COP</p>
-                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="h-5 w-5 cursor-pointer duration-150 hover:text-red-500" onClick={handleRemove}>
-                            <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-                        </svg>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-};
 
 const CartShop = () => {
     const { cart } = useCart();
@@ -124,7 +89,7 @@ const CartShop = () => {
                 <div className="mx-auto max-w-5xl justify-center px-6 md:flex md:space-x-6 xl:px-0">
                     <div className="rounded-lg md:w-2/3">
                         {cart.items.map((item) => (
-                            <CartItem key={item._id} item={item} />
+                            <CartItem key={item.id} item={item} />
                         ))}
                     </div>
                     {cart.items.length > 0 ? (
